requirements.txt
pyyaml
spotipy
requests
python-jwt
click
aiohttp
beautifulsoup4